Observables

List of built-in observables. They take a TumorConfiguration as the first and often only argument.

GrowthDynamics.TumorObservables.allele_fractions

GrowthDynamics.TumorObservables.allele_size

GrowthDynamics.TumorObservables.allele_spectrum

GrowthDynamics.TumorObservables.common_snps

GrowthDynamics.TumorObservables.explode_into_shells

GrowthDynamics.TumorObservables.living_ancestor

GrowthDynamics.TumorObservables.mean_pairwise

GrowthDynamics.TumorObservables.mrca

GrowthDynamics.TumorObservables.mrca

GrowthDynamics.TumorObservables.npolymorphisms

GrowthDynamics.TumorObservables.pairwise

GrowthDynamics.TumorObservables.pairwise

GrowthDynamics.TumorObservables.polymorphisms

GrowthDynamics.TumorObservables.popsize_on_shells

GrowthDynamics.TumorObservables.population_size

GrowthDynamics.TumorObservables.positions

GrowthDynamics.TumorObservables.sampled_allele_fractions

GrowthDynamics.TumorObservables.tajimasd

GrowthDynamics.TumorObservables.total_population_size

GrowthDynamics.TumorObservables.allele_fractions — Function

Dictionary of (SNP, freq).

GrowthDynamics.TumorObservables.allele_size — Function

Dictionary (SNP, population count)

GrowthDynamics.TumorObservables.allele_spectrum — Method

allele_spectrum(state;[ threshold=0.0, read_depth=total_population_size(state)])

Return a DataFrame with count, frequency of every polymorphism. Additionally sample from the population.

GrowthDynamics.TumorObservables.common_snps — Method

common_snps(::TumorConfiguration; filterdead=true)

List polymorphisms that are common to all genotypes.

Optional arguments:

filterdead=true: exclude unpopulated genotypes?

GrowthDynamics.TumorObservables.explode_into_shells — Method

explode_into_shells(v, o, a; a0=)

Take a vector of cartesian coordinates v, center them around the midpoint o, and return a dictionary radius=>coordinates where a0<= radius <= max(||v||) in increments of a.

GrowthDynamics.TumorObservables.living_ancestor — Method

Earliest living ancestor.

GrowthDynamics.TumorObservables.mean_pairwise — Method

Diversity (mean pairwise difference of mutations) of a population.

GrowthDynamics.TumorObservables.mrca — Method

Return index of the most recent common ancestor between (i,j) in a phylogeny.

GrowthDynamics.TumorObservables.mrca — Method

Return index of the most recent common ancestor in a phylogeny.

GrowthDynamics.TumorObservables.npolymorphisms — Method

npolymorphisms(S::TumorConfiguration)

Number of polymrphisms

GrowthDynamics.TumorObservables.pairwise — Method

pairwise(S::TumorConfiguration, i, j)

Number of pairwise genomic differences between genotype indices i,j.

GrowthDynamics.TumorObservables.pairwise — Method

pairwise(S::TumorConfiguration)

Matrix of pairwise differences. filterdead: Do not include extinct genotypes.

GrowthDynamics.TumorObservables.polymorphisms — Method

polymorphisms(S::TumorConfiguration)

Vector of polymorphisms (segregating sites).

GrowthDynamics.TumorObservables.popsize_on_shells — Method

popsize_on_shells(T, outer, [o=midpoint(T.lattice)])

Creates a dictionary genotype => trajectory where trajectory is a vector of population size. trajectory[r] is the the population size on a L2-norm shell of radius r around o.

r ranges between 1..outer in increments of a (defaults to the lattice spacing).

Set deleteone/zero=false to keep the wildtype/count empty sites.

GrowthDynamics.TumorObservables.population_size — Method

Dictionary (genotype, population size)

GrowthDynamics.TumorObservables.positions — Method

positions(state, g)

Returns coordinates of cells of genotype g.

GrowthDynamics.TumorObservables.sampled_allele_fractions — Function

sampled_allele_fractions(S::TumorConfiguration[, t=0, samples=length(S.meta.npops)])

Randomly sample genotypes(!) and calculate frequencies of contained SNPs. Return a dictionary (SNP, freq).

GrowthDynamics.TumorObservables.tajimasd — Method

See https://en.wikipedia.org/wiki/Tajima%27s_D

GrowthDynamics.TumorObservables.total_population_size — Method

Total population size. Duh.