States

A state is a collection of

either a lattice<:RealLattice if the model is spatial (see lattices), or a dummy placeholder NoLattice if it is not.

a phylogenetic tree, which is a directed graph where the root(s) are the wildtype(s).

meta data about the population and its genetics. See Meta Data.

the time the state has been evolved for by invoking dynamics on it.

States are of type TumorConfiguration.

GrowthDynamics.TumorConfigurations.TumorConfiguration — Type

TumorConfiguration(lattice, [phylogeny])

Wraps an existing lattice in a TumorConfiguration. Calculates meta.npops automatically.

If phylogeny is not given, it defaults to an empty graph.

It is normally not required to manipulate meta data, phylogeny or lattice directly. Convenient methods to manipulate state are provided.

julia>     state, _ = uniform(CubicLattice, 32; g=0)(GrowthDynamics.Lattices.CubicLattice{Int64, Array{Int64, 3}}
0	genotypes
0	population, nothing)
julia>     # fails because the new genotype is yet unknown
           state[16,16,16] = 1ERROR: ArgumentError: Genotype 1 is not know. Use push!(::TumorConfiguration, 1) first.
julia>     # make it known first
           push!(state, 1)1
julia>     state[16,16,16] = 1;
julia>     state.meta1-element GrowthDynamics.TumorConfigurations.MetaData{Int64}:
 (genotype = 1, npop = 1, fitness = 1.0, snps = nothing, age = (0, 0.0))
julia>     state[15:17,15:17,15:17]3×3×3 Array{Int64, 3}:
[:, :, 1] =
 0  0  0
 0  0  0
 0  0  0

[:, :, 2] =
 0  0  0
 0  1  0
 0  0  0

[:, :, 3] =
 0  0  0
 0  0  0
 0  0  0
julia>     push!(state, 2)2
julia>     state[1:size(state,1), 1, 1] .= 2;
julia>     state.meta2-element GrowthDynamics.TumorConfigurations.MetaData{Int64}:
 (genotype = 1, npop = 1, fitness = 1.0, snps = nothing, age = (0, 0.0))
 (genotype = 2, npop = 33, fitness = 1.0, snps = nothing, age = (0, 0.0))

Convenience constructors

Some convenience methods for common geometries are provided

GrowthDynamics.TumorConfigurations.uniform — Function

uniform(::Type{T<:RealLattice}, L; g=0, a=1.0)

Return a configuration on a lattice of type T with linear extension L, filled with genotype g.

Example

julia> uniform(HexagonalLattice, 128; g=1)
(HexagonalLattice{Int64, Matrix{Int64}}
1	genotypes
16641	population, nothing)

GrowthDynamics.TumorConfigurations.spheref — Function

spheref(::Type{T}, L::Int; r = 0, g1=0, g2=1) where LT<:Lattices.RealLattice

Fill lattice of type T (e.g CubicLattice) with genotype g1 and put a (L2-)sphere with genotype g2 that occupies approx. a fraction f of the lattice at the center.

GrowthDynamics.TumorConfigurations.spherer — Function

spherer(::Type{T}, L::Int; r = 0, g1=0, g2=1) where LT<:Lattices.RealLattice

Fill lattice of type T (e.g CubicLattice) with genotype g1 and put a (L2-)sphere of approx. radius r with genotype g2 at the center.

GrowthDynamics.TumorConfigurations.single_center — Function

single_center(::Type{T<:RealLattice}, L; g1=1, g2=2)

Put a single cell of genotype g2 at the midpoint of a lattice filled with g1.

GrowthDynamics.TumorConfigurations.half_space — Function

half_space(::Type{T<:RealLattice}, L; g1=1, g2=2)

Instantiates a population of genotype g1 on a lattice of type T. Fills the last dimension with g2 up to fraction f

Example

julia> using GrowthDynamics.TumorConfigurations

julia> state = half_space(CubicLattice, 32, f=1/4, g1=1, g2=2)[1]
CubicLattice{Int64, Array{Int64, 3}}
2	genotypes
35937	population

julia> state.meta
2-element MetaData{Int64}:
 (genotype = 1, npop = 27225, fitness = 1.0, snps = nothing, age = (0, 0.0))
 (genotype = 2, npop = 8712, fitness = 1.0, snps = nothing, age = (0, 0.0))

GrowthDynamics.TumorConfigurations.sphere_with_single_mutant_on_outer_shell — Function

sphere_with_single_mutant_on_outer_shell(::Type{<:RealLattice}, L::Int; r, s=1.0)

Fill a ball of radius r with genotype 1. Place a single cell of genotype 2 on the outermost shell at random.