Available dynamics

Routines to evolve a given state for a number of timesteps.

All routines take a state::TumorConfiguration as their first argument, and further parameters like mutation rate as keyword arguments.

GrowthDynamics.LatticeTumorDynamics.moran! — Function

moran!(state::NoLattice{Int}; <keyword arguments>)

(Extended) Moran dynamics on an unstructured population. Grow until carrying capacity is reach. After that individuals begin replacing each other.

Arguments

T::Int: the number of steps to advance.

fitness: function that assigns a fitness value to a genotype g::Int.

p_grow=1.0: Probability with which to actually proliferate. If no proliferation happens, mutation might still occur.

mu: mutation rate.

mu_type=[:poisson, :fixed]: Number of mutations is fixed, or Poisson-distributed.

genome_length=10^9: Length of the haploid genome.

d: death rate.

baserate: progressing real time is measured in 1/baserate.

prune_period: prune the phylogeny periodically after no. of steps.

prune_on_exit: prune before leaving the simulation loop.

callback: function of state and time to be called at each iteration. Used primarily for collecting observables during the run.

abort: condition on state and time under which to end the run.

GrowthDynamics.LatticeTumorDynamics.eden_with_density! — Function

eden_with_density!(state::RealLattice{Int}; <keyword arguments>)

Moran-like dynamics on an spatially structured population. Each step is either a death or (potential) birth and mutation event.

Individuals die at a rate d. Birthrate depends linearily on the number of neighbors.

Arguments

T::Int: the number of steps to advance.

fitness: function that assigns a fitness value to a genotype. Takes arguments (state, old genotype, new_genotype).

p_grow=1.0: Probability with which to actually proliferate. If no proliferation happens, mutation might still occur.

mu=0.0: mutation rate.

mu_type=[:poisson, :fixed]: Number of mutations is fixed, or Poisson-distributed.

genome_length=10^9: Length of the haploid genome.

d=0.0: death rate. Zero halts the dynamics after carrying capacity is reached.

baserate=1.0: progressing real time is measured in 1/baserate.

prune_period=0: prune the phylogeny periodically after no. of steps.

prune_on_exit=true: prune before leaving the simulation loop.

callback: function of state and time to be called at each iteration. Used primarily for collecting observables during the run.

abort: condition on state and time under which to end the run.

GrowthDynamics.LatticeTumorDynamics.exponential! — Function

exponential!(state::NoLattice{Int}; <keyword arguments>)

Run exponential growth on an unstructered population.

Arguments

T::Int: the number of steps (generations) to advance.

fitness: function that assigns a fitness value to a genotype g::Int.

mu: mutation rate.

baserate: progressing real time is measured in 1/baserate.

prune_period: prune the phylogeny periodically after no. of steps.

prune_on_exit: prune before leaving the simulation loop.

callback: function of state and time to be called at each iteration. Used primarily for collecting observables during the run.

abort: condition on state and time under which to end the run.