Meta Data

Every TumorConfiguration has MetaData attached. They contain for every genotype

A population count. While in principle redundant, it is significantly cheaper than enumerating the lattice.

A fitness value

TODO: A death rate

Mutations: Either nothing, or a vector of Int if mutations are present.

A timestamp when the genotype entered the population.

Additionally, meta.misc is a dictionary for storing arbitrary user-defined key-value pairs.

Convenient getter and setter routines are provided. Let's demonstrate with an example similar to the one from the Quick Start section, but label genotypes with strings of fixed length (from InlineStrings.jl)

julia> using InlineStrings
julia> import Base: zero
julia> zero(::Type{String7}) = String7("00-00")zero (generic function with 62 methods)
julia> state, _ = spheref(HexagonalLattice, 128, f=1/10, g1=String7("00-00"), g2=String7("00-AA"))(GrowthDynamics.Lattices.HexagonalLattice{InlineStrings.String7, Matrix{InlineStrings.String7}}
1	genotypes
1921	population, CartesianIndex{2}[CartesianIndex(60, 74), CartesianIndex(62, 74), CartesianIndex(64, 74), CartesianIndex(65, 74), CartesianIndex(66, 74), CartesianIndex(68, 74), CartesianIndex(70, 74), CartesianIndex(56, 75), CartesianIndex(58, 75), CartesianIndex(59, 75)  …  CartesianIndex(63, 119), CartesianIndex(64, 119), CartesianIndex(65, 119), CartesianIndex(66, 119), CartesianIndex(67, 119), CartesianIndex(68, 119), CartesianIndex(69, 119), CartesianIndex(70, 119), CartesianIndex(71, 119), CartesianIndex(65, 120)])
julia> eden_with_density!(state;
         label=(state,g)->String7( join(rand('0':'9',2))*"-"*join(rand('A':'Z',2)) ),
         T=1024, # timesteps
         mu=1e0,  # mutation rate per genome (not site!)
         d=1/100, # death rate
         fitness=(s,g_old,g_new)->1.0 + 0.1*randn() # function to assign fitness to new genotypes
       )

julia> state.meta546-element GrowthDynamics.TumorConfigurations.MetaData{InlineStrings.String7}:
 (genotype = InlineStrings.String7("00-AA"), npop = 1980, fitness = 1.0, snps = nothing, age = (0, 0.0))
 (genotype = InlineStrings.String7("46-YQ"), npop = 1, fitness = 0.8026922162866656, snps = [110084262, 253811787, 491383196, 565145360, 749194060], age = (0, 0.0))
 (genotype = InlineStrings.String7("83-TG"), npop = 3, fitness = 1.0835397026515496, snps = [293904862], age = (2, 0.01668181987508624))
 (genotype = InlineStrings.String7("96-XJ"), npop = 1, fitness = 0.869823904746547, snps = [314518933, 400909946], age = (5, 0.09349040835582655))
 (genotype = InlineStrings.String7("17-SF"), npop = 2, fitness = 0.9834021327434191, snps = [435520975], age = (6, 0.09505816477653363))
 (genotype = InlineStrings.String7("88-XB"), npop = 1, fitness = 1.1087871348049845, snps = [479112480, 943209862], age = (9, 0.13756267168600303))
 (genotype = InlineStrings.String7("71-ST"), npop = 3, fitness = 0.9218723537854466, snps = [435520975, 517628781, 954631554], age = (14, 0.1967250402915523))
 (genotype = InlineStrings.String7("54-ZW"), npop = 3, fitness = 0.8723606893869594, snps = [238562147], age = (19, 0.23117396974922502))
 (genotype = InlineStrings.String7("91-BB"), npop = 1, fitness = 1.1097052357136297, snps = [205402397, 520761909, 569074237, 874639413], age = (25, 0.28479566778859683))
 (genotype = InlineStrings.String7("73-LC"), npop = 2, fitness = 0.9429685908671765, snps = [360126036], age = (29, 0.31304712616825114))
 ⋮
 (genotype = InlineStrings.String7("07-RL"), npop = 1, fitness = 0.8114989887499285, snps = [83928989, 217679427, 357276393, 358069069, 359573704, 433961045, 448115477, 605422370, 729929773, 767131980, 862382472, 933138097], age = (1007, 8.747743410992854))
 (genotype = InlineStrings.String7("51-YI"), npop = 1, fitness = 0.9337429607126253, snps = [13850082, 14460264, 200586632, 244444181, 293904862, 365308675, 489227644, 516780669, 517673178, 845899473, 942147710], age = (1008, 8.750906005540417))
 (genotype = InlineStrings.String7("54-VQ"), npop = 1, fitness = 1.0028881748526348, snps = [13850082, 244444181, 293904862, 352518543, 489227644, 516780669, 845899473, 942147710], age = (1009, 8.754161159938796))
 (genotype = InlineStrings.String7("80-VJ"), npop = 1, fitness = 0.9319402697594115, snps = [136483025, 225287006, 338403002, 523519553, 716980061, 757821416, 777732472, 841752911], age = (1012, 8.79058190767328))
 (genotype = InlineStrings.String7("40-DK"), npop = 1, fitness = 0.7055970524742294, snps = [90080957, 263933125, 290799940, 396396699, 562957186, 598192109, 793808883], age = (1013, 8.795383238163371))
 (genotype = InlineStrings.String7("17-HB"), npop = 1, fitness = 0.9388069469251132, snps = [245073204, 603146202, 627677457, 745372988, 827285863, 839893191, 974079460], age = (1014, 8.798888029075732))
 (genotype = InlineStrings.String7("84-YB"), npop = 1, fitness = 1.064848627436434, snps = [26024992, 90080957, 263933125, 396396699, 488920527, 523992051, 562957186, 715258365, 755639877, 775794476, 793808883, 955622173], age = (1016, 8.820333546404333))
 (genotype = InlineStrings.String7("59-QT"), npop = 1, fitness = 1.1537076301393268, snps = [280633622, 359841454, 435300347, 518237875, 520037655, 537881028, 689029275], age = (1017, 8.820650346016288))
 (genotype = InlineStrings.String7("91-XL"), npop = 1, fitness = 0.7272769035891286, snps = [11410131, 159432085, 358283337, 475374470, 482980351, 626247857, 675662536, 760970614, 907153855], age = (1024, 8.872836826777796))

We can query according to index

julia> state.meta[2, :genotype]"46-YQ"

julia> state.meta[5:10, :npop]6-element Vector{Int64}:
 2
 1
 3
 3
 1
 2

or genotype

julia> state.meta[g="46-YQ"](genotype = InlineStrings.String7("46-YQ"), npop = 1, fitness = 0.8026922162866656, snps = [110084262, 253811787, 491383196, 565145360, 749194060], age = (0, 0.0))

julia> state.meta[g="46-YQ", :fitness] = 1.11.1

Note

In places where performance is paramount, getter and setter should be called like state.meta[2, Val(:genotype)] to circumvent dynamic dispatch.

Alternatively, getgenotype(state, id), setgenotype!(state, id), etc. are provided.

Because MetaData implements Julia's array interface, iterating is supported, and because each MetaDatum is a NamedTuple, we can for example convert to a DataFrame

julia> using DataFrames
julia> filter(v->v.npop>10, state.meta) |> DataFrame2×5 DataFrame
 Row │ genotype  npop   fitness   snps                             age         ⋯
     │ String7   Int64  Float64   Union…                           Tuple…      ⋯
─────┼──────────────────────────────────────────────────────────────────────────
   1 │ 00-AA      1980  1.0                                        (0, 0.0)    ⋯
   2 │ 84-ZM        11  0.866837  [7944355, 664146836, 865781806]  (227, 2.644
                                                                1 column omitted

API

GrowthDynamics.TumorConfigurations.MetaData — Type

MetaData{T}

Indexable structure to store information about a population.

Access

Use meta[id, :field] or meta[g=G, :field] to access and set information for genotype G, or the genotype with id id respectively.

id is also the number of the vertex in the phylogeny corresponding to a given genotype.

:field is one of

:npop - population size

:fitness - fitness value

:genotype - genotype corresponding to a given id

:snps: - vector of integers representing mutations a genotype carries. nothing if no mutations are present.

:age: tuple of (timestep, realtime) of when the genotype was first instantiated.

Useful for putting lengths on the branches of a phylogeny.

Additionally, a global field meta.misc::Dict{Any,Any} exists to store arbitrary, user-defined information.

See also MetaDatum

GrowthDynamics.TumorConfigurations.MetaDatum — Type

MetaDatum{T,S}

NamedTuple to store information about a single genotype of type T that carries mutations of type S (default Int).

Missing docstring.

Missing docstring for index(::MetaData{T}, ::T) where {T}. Check Documenter's build log for details.

GrowthDynamics.TumorConfigurations.index! — Function

index!(::MetaData)

Reindex the metadata.